casting
Former-commit-id: 780e07dfc04e2e8fea4108dd48431022acc88cc9

+++ modules/EngineLayer/QueryManager/Server.cpp
@@ -187,7 +187,7 @@ void QueryManagerServer::Query(Query_Result& _return, const SessionID sessionID,
       }
       std::string type = m_compression.getCompressionTypeString();
       if ( lvl)
-	type += " " + std::to_string( lvl);
+	type += " " + std::to_string( ( long long int)lvl);
       double percent = 100.0 * ( double)( compressed_result->size() - 12) / ( double)num_bytes;
       LOGGER << "        to " << GetNiceSizeString( compressed_result->size() - 12) // header
 	     << " from " << GetNiceSizeString( num_bytes) << " ( " << percent << "% "
